Add php 8.2 to test pipeline (#32)
* Add php 8.2 to test pipeline #30
* Ignore unmatched errors in phpstan
Since we support a wide range of versions for upstream packages,
and some type issues have been fixed in newer versions,
errors may or may not exist.
* Remove composer.lock from version control
Some peer dependencies don't work with php 8.0 and php 8.2.
The problem is that they are defined in the lockfile. Instead of
fixing this during the pipeline run (as it is too specific), we just
ignore the composer.lock file and let composer manage it itself.
* Pin phpstan and phpcs to specific versions
The reason is that they might break the pipeline when we download
the latest minor release, as they might adjust their rulesets.
This was the main reason for committing lockfile in the past.